refactor(reviews and campground): restructured the code for reviews and campgrounds to leave index.js cleaner
